pyproject.toml
[tool.poetry]
name = "aepp-sdk"
version = "3.1.0"
description = "Python SDK to interact with the Æternity blockchain"
authors = [
  "Andrea Giacobino <no.andrea@gmail.com>", 
  "Shubhendu Shekhar <shubhendu@apeunit.com>"
]
license = "ISC"
homepage = "https://github.com/aeternity/aepp-sdk-python"
repository = "https://github.com/aeternity/aepp-sdk-python"
readme = "README.md"
keywords = ["aeternity", "blockchain", "sdk"]
include = ["CHANGELOG.md"]
classifiers = [
    'License :: OSI Approved :: ISC License (ISCL)',
    'Intended Audience :: Developers',
    'Operating System :: OS Independent',
]
packages = [
    { include = "aeternity" },
]

[tool.poetry.dependencies]
python = "^3.6"
base58 = "^0.2.5"
click = "^7.0"
rlp = "^0.6.0"
PyNaCl = "^1.3"
requests = "^2.20"
validators = "^0.12.6"
semver = "^2.8"
namedtupled = "^0.3.3"
Deprecated = "^1.2"
websockets = "^7.0"

[tool.poetry.dev-dependencies]
pytest = "^4.3"
flake8 = "^3.7"

[tool.poetry.scripts]
aecli = "aeternity.__main__:run"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"


